+++ dffml/source/dfpreprocess.py
@@ -75,7 +75,7 @@ class DataFlowPreprocessSourceContext(BaseSourceContext):
                         key=key, data=data,
                     )
 
-    async def __aenter__(self) -> "DataFlowSourceContext":
+    async def __aenter__(self) -> "DataFlowOldSourceContext":
         self.sctx = await self.parent.source().__aenter__()
 
         if isinstance(self.parent.config.dataflow, str):