+++ CHANGELOG.md
@@ -59,7 +59,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Example usage of sklearn operations
 - Example Flower17 species image classification
 - Configloading ablity from CLI using "@" before filename
-- Docstrings and doctestable example for DataFlowSource
+- Docstrings and doctestable example for DataFlowOldSource
 - XGBoost Regression Model
 - Pre-Trained PyTorch torchvision Models
 - Spacy model for NER
@@ -71,7 +71,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Operations plugin for image preprocessing
 - `-pretty` flag to `list records` and `predict` commands
 - daal4py based linear regression model
-- DataFlowSource can take a config file as dataflow via the CLI.
+- DataFlowOldSource can take a config file as dataflow via the CLI.
 - Support for link on conditions in dataflow diagrams
 - `edit all` command to edit records in bulk
 - Support for Tensorflow 2.2